Fix build failure with gnutls 3.0.
This is the line causing the issue (in inc/encrypt.hpp):
  typedef gnutls_connection_end gnutls_connection_end_t;

And it turns out that the new gnutls has compatibility #defines for both
gnutls_connection_end and gnutls_connection_end_t, so it breaks stuff
when expanding this line.

I'd guess it should be rather safe now to simply not redefine the type,
and just use gnutls_connection_end_t directly (see patch).